Improved release note tools - mk_esr_notes
Improved version of gather_notes.py. Figures out all of the version stuff from
product-details and version control.

It will need to be updated to deal with esr91 when the time comes, but for now
it will work.

A pretty much ready to go YAML file is generated based on the notes found in
previous betas.

The notes themselves will be split into new, changed, and fixed. But within those
categories they will need to be rearranged.