Update 115.4.1.yml
Fixed "no" typo.

I also tried to rewrite the summary for bug 1857196, as these lines are usually written to describe the problem that was fixed. So "Fixed: Calendar ICS parser improvements" makes it sound like "Calendar ICS parser improvements" was a problem that needed fixing!

I'm not sure my new wording is ideal, so feel free to modify it, but hopefully you get the gist of what I'm trying to do. Thanks :)